+++ packages/@react-native-windows/cli/src/runWindows/utils/commandWithProgress.ts
@@ -8,7 +8,12 @@ import {spawn, SpawnOptions} from 'child_process';
 import ora from 'ora';
 import spinners from 'cli-spinners';
 import chalk from 'chalk';
-import {CodedError, CodedErrorType} from '@react-native-windows/telemetry';
+import {
+  Telemetry,
+  CodedError,
+  CodedErrors,
+  CodedErrorType,
+} from '@react-native-windows/telemetry';
 
 function setSpinnerText(spinner: ora.Ora, prefix: string, text: string) {
   text = prefix + spinnerString(text);
@@ -153,4 +158,40 @@ export function newSuccess(text: string) {
 
 export function newInfo(text: string) {
   newSpinner(text).info(text);
+}
+
+/**
+ * Sets the process exit code and offers some information at the end of a CLI command.
+ * @param loggingIsEnabled Is verbose logging enabled.
+ * @param error The error caught during the process, if any.
+ */
+export function setExitProcessWithError(
+  loggingIsEnabled?: boolean,
+  error?: Error,
+): void {
+  if (error) {
+    const errorType =
+      error instanceof CodedError ? (error as CodedError).type : 'Unknown';
+
+    process.exitCode = CodedErrors[errorType];
+
+    if (loggingIsEnabled) {
+      console.log(
+        `Command failed with error ${chalk.bold(errorType)}: ${error.message}`,
+      );
+      if (Telemetry.isEnabled()) {
+        console.log(
+          `Your telemetry sessionId was ${chalk.bold(
+            Telemetry.getSessionId(),
+          )}`,
+        );
+      }
+    } else {
+      console.log(
+        `Command failed. Re-run the command with ${chalk.bold(
+          '--logging',
+        )} for more information.`,
+      );
+    }
+  }
 }